code restructure with commenting
